backend/requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.27.1


python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1


langchain==0.1.14
langchain-core==0.1.38
langchain-community==0.0.29


langchain-groq==0.1.3
groq==0.4.2


requests==2.31.0


tenacity==8.2.3
typing-extensions==4.10.0


orjson==3.9.15
python-multipart==0.0.9
